# packages/shared-types/src/types/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pinaka API",
    "version": "1.0.0",
    "description": "API documentation generated from canonical schema registry"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/invitations": {
      "get": {
        "summary": "List Invitations",
        "tags": [
          "Invitation"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Invitations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Invitation",
        "tags": [
          "Invitation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{id}": {
      "get": {
        "summary": "Get Invitation by ID",
        "tags": [
          "Invitation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Invitation",
        "tags": [
          "Invitation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Invitation",
        "tags": [
          "Invitation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Invitation deleted"
          }
        }
      }
    },
...
    "/api/v1/landlords": {
      "get": {
        "summary": "List Landlords",
        "tags": [
          "Landlord"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Landlords",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandlordListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Landlord",
        "tags": [
          "Landlord"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandlordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Landlord created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandlordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/landlords/{id}": {
      "get": {
        "summary": "Get Landlord by ID",
        "tags": [
          "Landlord"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Landlord details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandlordResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Landlord",
        "tags": [
          "Landlord"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandlordUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Landlord updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandlordResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Landlord",
        "tags": [
          "Landlord"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Landlord deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "InvitationCreate": {
        "type": "object",
        "description": "Schema for creating a Invitation"
      },
      "InvitationUpdate": {
        "type": "object",
        "description": "Schema for updating a Invitation"
      },
      "InvitationResponse": {
        "type": "object",
        "description": "Response schema for Invitation"
      },
...
      "LandlordCreate": {
        "type": "object",
        "description": "Schema for creating a Landlord"
      },
      "LandlordUpdate": {
        "type": "object",
        "description": "Schema for updating a Landlord"
      },
      "LandlordResponse": {
        "type": "object",
        "description": "Response schema for Landlord"
      }
    }
  },
...
}